Fixed typo for names of test, renamed OtherSuite.java
