+++ lib/build.gradle
@@ -80,8 +80,8 @@ task testRestSuite(type: Test) {
   outputs.upToDateWhen { false }
 }
 
-task testOtherSuite(type: Test) {
-  include '**/OtherSuite.class'
+task testUtilsSuite(type: Test) {
+  include '**/UtilsSuite.class'
   exclude '**/*Test.class'
   beforeTest { descriptor ->
     logger.lifecycle("Running test: " + descriptor)
